Replace depedency on custom facts
Hardware inventory used system_vendor and system_model which were custom
facts, not present in puppet/facter stock distribution. Facter, however,
ships with manufacturer/productname, which are the exact equivalent, so
use those instead.

While at it, get rid of board S/N from the hardware inventory; there's a
stock fact for this as well, but that S/N is useless anyway, so no
reason to have it take up space.