Fact query: adjust fact exclude filter
We currently were excluding macaddress/ipaddress/netmask, presumably to
avoid listing hundreds of virtual interfaces for VMs. Do the same for
ipaddress6 and network, as to avoid those as well and be consistent.

While at it, filter everything that starts with "---", i.e. Ruby
serialized objects that puppet puts in the database (ugh).