Make overview's problematic hosts a bit prettier
Overview's problematic host is currently lying: it says "have not
contacted the puppetmaster for the past hour", while the one-hour
duration has been configurable in settings.py for quite some time.

Adjust that to derive this from the setting. While at it, also make the
last update field a "time since" instead of the timestamp of the last
update, to make it more readable by humans.