Performance optimizations on the update views
Currently the hostlist view was making N queries, with N being the
number of hosts; instead of repeatedly calling update_set, use Django's
aggregation functions and make one single cheap query to get the list.

Similarly for packagelist, although the queries were as many as the
number of pending packages.