Apply tablesorter & odd/even colors on most tables
Tablesorter is useful in most of the views, however it was only applied
to a couple of them. Apply it in almost every view or, wheh it doesn't
make sense, apply a cycle of odd/even coloring of rows.

Since we apply this in almost every view, include tablesorter in the
base template and add the initial jQuery there for a global class
"sortable".

And while we're at it adding <thead> and <tbody> everywhere, (re)indent
tables to make them more readable.