Use MEDIA_URL rather than hard-coded /static/
Use MEDIA_URL instead of a hard-coded /static in templates; helps if
servermon doesn't run on root.

Since to reference MEDIA_URL from templates you need the RequestContext
and since this is quite ugly, use Django 1.3's render() function that
does this automatically. Additionally, to keep compatibility with 1.2,
add a "compat" module (enabled automatically) that includes a copy of
1.3's render() and loads it conditionally.

hwdoc already had this, without the render() hack; don't touch it for
now.