Unit Tests for servermon wide views
There are some views that are not installed in an app but are project
wide. These reside in file views.py of top level directory. It is not
possible for the unit testing framework to perform tests on that level
so unit tests for that are included in updates app