Add permission to change comments only in admin
This commit adds a Django permission that enables giving access to a
user to change only the comments fields of an Equipment model instance
of hwdoc. To use, a new user needs to be created, is_staff needs to be
True and the permission can_change_comment needs to be given to the
user. This allows workflows where a non-privileged user is capable of
updating that field and superusers can take over that and proceed with
whatever procecure has been put in place. Special care has been taken to
minimize info leaked to the new user