Populate equipment info from a ticketing system
This commit is a first effort to allow linking to a ticketing system. At
this point it is completely dumb. It works by having a person fill the
comments field of an equipment with string matching a regular expression
pattern configured in settings.TICKETING_URL and parses it on rutime
making the info available on a tickets attribute of the instance