Fix design error in hwdoc null=True, blank=True
This commit fixes a design error in hwdoc where CharFields and
TextFields where given both null=True and blank=True causing a case
where the absence of an entry could be declared by either a NULL value
or the empty string ''. We fix this by submitting one data migration
which translates all NULLs to empty strings, one schema migration to
change the models and the changed models themselves.