Re-implement the indicative statuses in hwdoc
There used to be an indicative status for an equipment in hwdoc results
which displayed whether there were any tickets of warnings for it. This
was no longer functioning when hwdoc was moved to bootstrap. This commit
cleans up the old CSS classes and utilizes bootstrap CSS to display the
info