Fix bug with rack search not matching in hwdoc
When Racks finally got names the search function was not updated and
searches for rack who would not match the rack id would fail. This
commit fixes that by looking for the specified search pattern in the
rack name. This does have the caveat of relying on the rack name to
contain the wanted number. We could also search in the position field
or RackPosition but that would probably confuse the users by producing
unexpected for them results