Search with rackunit notation would not always work
Due to the way rack/rackunit search was working it was possible to omit
results when searching. This has been circumvented by now using one
single place in the code where search takes place and adding some term
in the already rather long list. Had to also modify models a bit to
remove the default empty name for racks, the tests to not create racks
with empty names and any place where rack.pk was used instead of
rack.name. This closes #4118