Fix two bugs in hwdoc migrations
Backwards migration 0012 in hwdoc could fail in SQLite due to the
inexistence of an index. Fix it with an if using SQLite. This closes #26
Migration 0016 in hwdoc when run sequentially would create a content
type for Equipment model with no name attribute. Fix it by explicitly
providing a name in the migration. This closes #27