Adding key value pairs support, allow tags, attrs
The entire project is now capable of having key values pairs djangling
of any model instance. This is a new app and comes with its own models,
admin and tests.  The first client of this new support is the
Equipment model instance which adds a reverse relationship for easy
access and allows to easily attach key value pairs to equipment allowing
easy access to generalized operator added attributes like Asset Tags,
The second client of this is the Equipment Model model which allows for
the same kind of data in the model's giving the ability to add
attributes specific to a model like 'POE Ports: 1-8'. This closes #20