Display empty unit (Us) in racks
A number of changes to allow rack views to also show empty units so that
it can be easy to spot empty Us in a Rack. To do this we add a units
property to rack models, perform calculations in the rack view to create
virtual equipments for the empty units and change the template to not
cough up when it encounters those. Furthermore we add a units property
on equipment models and use it for the final calculations of displayed
rack units. This closes #24