Migrate old ticketing to new modular one
Have all view and templates use the new ticketing system. Finally fix
statistics in overview page to correctly display machines with tickets.
Datacenter page now is capable of display rackrows with tickets.
Adding the possibility to actively query a ticketing system for tickets
associated with a set of Equipments. This is implemented through a
django management command designed to be run via cron and given a set of
equipments (or all in the database) and query a ticketing system in
order to populate Ticket models. The system is modular in order to allow
for multiple Ticketing systems. Two ticketing system implementations are
provided. One dummy (doing nothing) and one depending on comments having
URLs matching an administrator specified pattern
Tests updated as well