ServerManagement made optional
Adding verbose_name and verbose_name_plural to all models was long due
anyway. Now we can also translate model names in admin. This refs #6
Since we are doing that, also add the designation "Optional" to
ServerManagement inline forms to inform users they do not have to fill
in those fields. This closes #39