Remove spaces from txt search results
The search view has the ability to return results in a very simplified
text format. This is a poor man's API. Should the project name have
spaces in it it makes the output rather difficult to handle as a column
based format. To avoid that, strip spaces from project names. This
should be sufficient. For more programmatic approaches, a proper REST
API is probably needed

This closes #54