Add a monkey_patch_command_execute function in compat
Add a new command that monkey patches the execute function for django
management commands. This is to avoid a django bug that was fixed in 1.5
but is present in previous versions and forces a sys.exit(1) on
CommandError. It is to be used only on unit tests where failure is a
success from the test's point of view.