Rename the useless urls.py.dist to urls.py
The idea behind urls.py.dist was to ensure administrators the installed
servermon could serve it under any URL they chose. The .dist suffix
would presumably clearly designate that this file needed to be copied
and messed with. In practice that has not been the case. So renaming it
to urls.py and modifying documentation accordingly. While it is still
possible to serve the project under a different URL that / by messing
around a bit with urls.py it is no longer the default