Racktables importer
Add a racktables to hwdoc import script. For now only mysql support
exists and we target version 0.20.4. Various assumptions had to be made
to make this possible. Unfortunately racktables and servermon vary quite
a lot in the way they structure their models and relations. It might
need updates depending on installation. We are obviously gonna accept
PRs though.

This closes #21