Allow for more than one lead characters in rack/u search
A quick way to search for a machine if you know its rack and u is
provided via the <rackname>U<unit_number> shortcut. This however allowed
for a very specific way to handle rack names which was <letter><two
digits>. Increase the number of letters allowed at the start. Make the U
or u lowercase character required as well. Update the tests to reflect
that.

This closes #86