Fix 0016_can_change_comment TransactionManagementError
Migration 16 causes TransactionManagementError under Django 1.6, a
previous effort to fix it was done in 58aec5f but it was only half
correct. Drop the anyway faulty get_or_create and replace it with a get
Fix the backwards migration as well. Fow the fun part now, we
erroneously had that get_or_create in the first place because we needed
the contenttypes instances before they were populated by the post_syncdb
signal. Force the contenttypes framework to update itself so we can use
the instances right away