Import django admin media
Since we need to support multiple django versions we can not rely on the
staticfiles app existence just yet. This will change in the future but
for now ship django's admin media in the project and update templates to
use the shipped media. This removes usage of admin_media_prefix which
was deprecated and no longer exists on Django 1.5 and up. The version of
admin media we import is 1.3, slightly patched. This closes #78