Wrap hwdoc_populate_tickets in a transaction
In #66, there was some concern about cron commands running while the
entire project is in maintenance mode and possibly DDL commands are run
on the database. To avoid such issues as well as avoiding
inconsistencies, make_updates has for a long time been wrapped in a
commin_on_success decorator. Wrap populate_tickets in the same decorator
as well. This closes #66