coverage: try to omit some files from coverage checks
Avoid reporting the coverage of some vendor specific plugins for bmc and
ticket modules. These are generally difficult to test right now and not
used very much so will skip them for now. Move the entire coverage
configuration into a .coveragerc file to make it clearer and reusable,
especially by the coveralls program