Update WSGI module
Changing the entire WSGI setup is warranted. Stop the practice of having
the user configure the path in the WSGI application module, and move it
along the rest of the project since it now is a shipped module. This
should allow easier setting up of the project in application servers.
To achieve this migrate to the Django 1.4 way of having the project
organized, including the settings-test.py module, the manage.py module
and urls and settings
Update the example gunicorn configuration to work with the WSGI way
since this is the prefered method since Django 1.5