Getting ready for a 0.7 release
Add support for Django 1.5 and Django 1.6 as tests pass for them and the
supplied gunicorn as well as the wsgi configuration works fine. In 1.7
these things no longer work and need to be updated. That and the
migrations changes in 1.7 warrant a non support status for servermon 0.7
Update global copyrights as well