Reorganize location information generation
Location generation information in the puppet hostview is suboptimal and
has some hardcoded values. It fails for any host not having any other
specified attributes resulting in a no location information line
displayed. Reorganize it in a anonymous function way, avoiding that
problem and move the link approach into an attribute. To get better
output out of the view, add permalink methods to rack and datacenter and
use them in the view. Also to get code coverage for the new code, add
extra tests