Display a tooltip for packages without an installed version
Packages without an installed version are packages that will be
installed if a safe-upgrade (aka dist-upgrade) is run on the host in
question. Display an informative tooltip for those packages to help the
user understand why they are listed. To ensure we have a testbed for
this, update the dev_fixtures generating command and update the
sampledata fixtures shipped with the distribution

This closes #157