travis support for mysql/postgres tests
Add travis test support for mysql/postgres tests in addition to the
standard sqlite tests. To achieve this, change tox.ini to also accept as
an environment variable the DB backend to test against and ship settings
files that allow testing against mysql and postgres. To allow seamless
importing between though, replace the dash (-) character with an
underscore in the file names