updates: Rewrite hostlist and packagelist
hostlist and packagelist views in updates app would abuse the SUM SQL
function to annotate the Package and Host querylist respectively with
security updates info. However at least PostgreSQL will not allow SUM
function to be run on BooleanFields causing Exception and 500 errors.
Rewrite that part of the views to allow fixing this