Force virtualenv version in travis builds
To bypass a number of issues with recent builds in travis, force the
virtualenv version to 13.1.2 as it is the last stable virtualenv version
we had no problems with. This is a temporary measure to get Travis
builds working fine again and needs to be investigated better and fixed.
The solution however seems to be dependent on migrating to SemVer and
pbr > 0.11 which is going to take a while to happen