Linkify rack and dc in equipment and search views
Per #195, it would be useful from a UX perspective, to have links
directly to rack from the equipment and the search results views. Adding
that as well as a link to the DC in the search results view. To make the
result look better and still not break backwards compatibility with the
display or the rackunit, add a new row containing just the rack

This closes #195