When DEBUG=True use static to serve staticfiles
We could be debugging in other environments than django's builtin
runserver and having static files served can be a big difference in
debugging time. Fix the here() lambda in order to serve the correct
directory for that to work fine. This is something that used to exist in
the past but was removed in 58ba9cfcf23d (which was a WIP and should not
have been merged as is)