Use an explicit relative import for _bmc_common
_bmc_common private module abstracts away some of the common code shared
by all hwdoc management commands. While a relative implicit import is
fine in python 2.x, in python 3.x it is not allowed. Migrate to an
explicit relative import. Using an absolute import here would break the
idea behind this and would expose the django management commands
hierarchy to the module which is not really useful or wanted