Use python_2_unicode_compatible in all models
Specifying the __unicode__() function in python 3 is effectively a noop
as it no longer has special treatment as it does in python 2. Instead
__str__ is meant to be used. However that function has a different
meaning in python 2 and Django is kind enough to provide a decorator
that handles that trasparently for us. Use that decorator wherever we
used to specify the __unicode__() function and convert all __unicode__
instances to __str__