Make sure we make encapsulate strings in an iterable
We used to rely on a not very future proof behavior of strings, that is
that they did not have the __iter__() function whereas still being
iterables. This is not python 3 compatible however given strings do have
an __iter__() function in python 3. Detect if it is a string we are
talking about and force the tuple