garbage collection host support
In many environments, active record is no longer used and people have
migrated to PuppetDB. Puppet 4 does not even support ActiveRecord
anymore. For these cases and until we support PuppetDB, do a garbage
collection of hosts. Piggyback in the make_updates command for that,
which is not architecturally the best option but is good enough for now